scripts/make_fit: Support an initial ramdisk
FIT (Flat Image Tree) allows a ramdisk to be included in each
configuration. Add support for this to the script.

This feature is not available via 'make image.fit' since the ramdisk
likely needs to be built separately anyway, e.g. using modules from
the kernel build. A later patch in this series provides support for
doing that.

Note that the uncompressed size is not correct when a ramdisk is use,
since it is too expensive to decompress the ramdisk.

Series-changes: 2
- Don't compress the ramdisk as it is already compressed

Series-changes: 3
- Add a comment at the top of the file about the -r option
- Count the ramdisk in the total files
- Update the commit message

Series-changes: 3
- Update the commit message

Signed-off-by: Simon Glass <sjg@chromium.org>