scripts/make_fit: Speed up operation
The kernel is likely at least 16MB so we may as well use that as a step
size when reallocating space for the FIT in memory. Pack the FIT at the
end, so there is no wasted space.

This reduces the time to pack by an order of magnitude, or so.

Series-changes: 3
- Move the ramdisk chunk into the correct patch

Signed-off-by: Simon Glass <sjg@chromium.org>